// src/modules/shared/enums/order-status.enum.ts
export enum OrderStatus {
  PENDING = 'PENDING',
  DELIVERED = 'DELIVERED',
  DELIVERING = 'DELIVERING',
  FILLED = 'FILLED',
  PROCESSING = 'PROCESSING',
  RECEIVED = 'RECEIVED',
  ON_WAY = 'ON_WAY',
  ACCEPTED = 'ACCEPTED',
  COMPLETED = 'COMPLETED',
  DECLINED = 'DECLINED',
  CANCELLED_BY_VENDOR = 'CANCELLED_BY_VENDOR',
  CANCELLED_BY_CUSTOMER = 'CANCELLED_BY_CUSTOMER',
  CANCELLED_BY_ADMIN = 'CANCELLED_BY_ADMIN',
}